+++ openpilot/selfdrive/ui/mici/layouts/settings/settings.py
@@ -7,6 +7,7 @@ from openpilot.selfdrive.ui.mici.layouts.settings.device import DeviceLayoutMici
 from openpilot.selfdrive.ui.mici.layouts.settings.developer import DeveloperLayoutMici
 from openpilot.selfdrive.ui.mici.layouts.settings.software import SoftwareLayoutMici
 from openpilot.selfdrive.ui.mici.layouts.settings.firehose import FirehoseLayout
+from openpilot.selfdrive.ui.mici.layouts.settings.stats import StatsLayoutMici
 from openpilot.system.ui.lib.application import gui_app, FontWeight
 
 
@@ -27,6 +28,10 @@ class SettingsLayout(NavScroller):
     theme_panel = ThemeLayoutMici()
     theme_btn = SettingsBigButton("theme", "", gui_app.texture("icons_mici/settings/theme.png", 64, 58))
     theme_btn.set_click_callback(lambda: gui_app.push_widget(theme_panel))
+
+    self._stats_layout = StatsLayoutMici()
+    stats_btn = SettingsBigButton("statistics", "", gui_app.texture("icons_mici/wheel.png", 64, 64))
+    stats_btn.set_click_callback(lambda: gui_app.push_widget(self._stats_layout))
 
     network_panel = NetworkLayoutMici()
     network_btn = SettingsBigButton("network", "", gui_app.texture("icons_mici/settings/network/wifi_strength_full.png", 76, 56))
@@ -51,6 +56,7 @@ class SettingsLayout(NavScroller):
     self._scroller.add_widgets([
       toggles_btn,
       theme_btn,
+      stats_btn,
       network_btn,
       device_btn,
       software_btn,
@@ -60,3 +66,7 @@ class SettingsLayout(NavScroller):
     ])
 
     self._font_medium = gui_app.font(FontWeight.MEDIUM)
+
+  @property
+  def stats_layout(self):
+    return self._stats_layout